clean ups of tests not to use dtype

+++ python/test_inmf.py
@@ -9,7 +9,7 @@ windowHop = 4
 
 d = ricaudio.INMF(8, 2, 10, 0.5, 15, 1e-17)
 
-a = scipy.zeros((14, 8), dtype = 'f4')
+a = scipy.zeros((14, 8))
 a[:4, 2] = 1
 a[5:9, 6] = 1
 a[10:, 2] = 1
